refactor: address PR review comments for timeline pinia migration [WTEL-9717](https://webitel.atlassian.net/browse/WTEL-9717)
