+++ src/modules/timeline/enums/TimelineEventType.ts
@@ -1,7 +1,11 @@
+// ContactsTimelineEventType is missing from the types barrel of '@webitel/api-services/gen'
+
+import { WebitelCasesCaseTimelineEventType } from '@webitel/api-services/gen';
+import { ContactsTimelineEventType } from '@webitel/api-services/gen/models';
+
 export const TimelineEventType = {
-	Chat: 'chat',
-	Call: 'call',
-	Email: 'email',
+	...ContactsTimelineEventType,
+	...WebitelCasesCaseTimelineEventType,
 } as const;
 
 export type TimelineEventType =
